requirements.txt
argparse>=1.2.1
arrow>=0.4.2
base32-crockford>=0.2.0
behave>=1.2.4
bunch>=1.0.1
click>=6.7
configobj>=5.0.5
coverage>=3.7.1
datadiff>=1.1.5
setuptools>=14.3.0
flake8>=2.1.0
jsonpointer>=1.3
mccabe<0.7.0,>=0.6.0
mock>=1.0.1
nose>=1.3.3
parse>=1.6.4
parse-type>=0.3.4
pep8>=1.5.7
pyflakes>=2.0.0
python-dateutil>=2.2
pyyaml
requests>=2.3.0
requests-testadapter>=0.3.0
six>=1.6.1
twine
